fastapi_base/imageGen_BG/service/pipeline_service.py
from diffusers import AutoPipelineForText2Image, StableDiffusionInpaintPipeline
from PIL import Image
from typing import List, Dict, Any
import torch

from ..core.config import settings

class PipelineService:
    def __init__(self):
        self.config = settings.config
        self._inpaint_pipe = None
        self._text2img_pipe = None
        self._ip_adapter = None
    
    async def run_inpainting(
        self,
        image: Image.Image,
        mask: Image.Image,
        prompt: str,
        config: Dict[str, Any]
    ) -> List[Image.Image]:
        """Inpainting 파이프라인 실행"""
        if not self._inpaint_pipe:
            self._inpaint_pipe = self._load_inpaint_pipeline()
        
        # 기존 run_inpainting 로직 적용
        pass
    
    async def generate_with_ip_adapter(
        self,
        product_image: Image.Image,
        prompt: str,
        config: Dict[str, Any]
    ) -> List[Image.Image]:
        """IP-Adapter를 통한 이미지 생성"""
        if not self._text2img_pipe:
            self._text2img_pipe = self._load_text2img_pipeline()
        if not self._ip_adapter:
            self._ip_adapter = self._load_ip_adapter()
        
        # 배경 생성 후 IP-Adapter 합성 로직
        pass