fastapi_base/requirements.txt
# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML 라이브러리 (호환 버전으로 통일)
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Hugging Face 생태계 (버전 호환성 맞춤)
transformers==4.36.0
diffusers==0.25.0
accelerate==0.25.0
huggingface-hub==0.19.4
tokenizers==0.15.0

# 이미지 처리
Pillow==10.0.1
opencv-python-headless==4.8.1.78
rembg==2.0.50

# OpenAI
openai==1.3.0

# 유틸리티
pydantic==2.5.0
pydantic-settings>=2.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
numpy==1.24.4

# 추가 AI 도구 (선택사항)
controlnet-aux==0.0.7
xformers==0.0.22.post7

# 개발 도구
pytest==7.4.3
black==23.11.0